Fixes #1945

+++ myExpenses/src/main/java/org/totschnig/myexpenses/provider/BaseTransactionProvider.kt
@@ -217,12 +217,12 @@ abstract class BaseTransactionProvider : ContentProvider() {
         }
     }
 
-    fun callerIsNotSyncAdapter(uri: Uri): Boolean {
-        return !uri.getBooleanQueryParameter(
-            TransactionProvider.QUERY_PARAMETER_CALLER_IS_SYNCADAPTER,
-            false
-        )
-    }
+    fun callerIsNotSyncAdapter(uri: Uri) = !callerIsSyncAdapter(uri)
+
+    fun callerIsSyncAdapter(uri: Uri) = uri.getBooleanQueryParameter(
+        TransactionProvider.QUERY_PARAMETER_CALLER_IS_SYNCADAPTER,
+        false
+    )
 
     fun deleteTransaction(db: SupportSQLiteDatabase, uri: Uri): Int {
         val segment = uri.pathSegments[1]
@@ -247,27 +247,28 @@ abstract class BaseTransactionProvider : ContentProvider() {
                 arrayOf(segment)
             )
             //we delete the transaction, its children and its transfer peer, and transfer peers of its children
-            val count = if (uri.getQueryParameter(TransactionProvider.QUERY_PARAMETER_MARK_VOID) == null) {
-                //we delete the parent separately, so that the changes trigger can correctly record the parent uuid
-                db.delete(
-                    TABLE_TRANSACTIONS,
-                    WHERE_DEPENDENT,
-                    arrayOf<String?>(segment, segment)
-                ) + db.delete(
-                    TABLE_TRANSACTIONS,
-                    WHERE_SELF_OR_PEER,
-                    arrayOf<String?>(segment, segment)
-                )
-            } else {
-                val v = ContentValues()
-                v.put(KEY_CR_STATUS, CrStatus.VOID.name)
-                db.update(
-                    TABLE_TRANSACTIONS,
-                    v,
-                    WHERE_SELF_OR_RELATED,
-                    arrayOf(segment, segment, segment)
-                )
-            }
+            val count =
+                if (uri.getQueryParameter(TransactionProvider.QUERY_PARAMETER_MARK_VOID) == null) {
+                    //we delete the parent separately, so that the changes trigger can correctly record the parent uuid
+                    db.delete(
+                        TABLE_TRANSACTIONS,
+                        WHERE_DEPENDENT,
+                        arrayOf<String?>(segment, segment)
+                    ) + db.delete(
+                        TABLE_TRANSACTIONS,
+                        WHERE_SELF_OR_PEER,
+                        arrayOf<String?>(segment, segment)
+                    )
+                } else {
+                    val v = ContentValues()
+                    v.put(KEY_CR_STATUS, CrStatus.VOID.name)
+                    db.update(
+                        TABLE_TRANSACTIONS,
+                        v,
+                        WHERE_SELF_OR_RELATED,
+                        arrayOf(segment, segment, segment)
+                    )
+                }
             db.setTransactionSuccessful()
             count
         } finally {
@@ -1475,7 +1476,8 @@ abstract class BaseTransactionProvider : ContentProvider() {
         val accountQuery = uri.accountSelector
 
         //Grand total account or aggregate account for type or flag
-        val forHome: String? = if (uri.getQueryParameter(KEY_ACCOUNTID) == null && uri.getQueryParameter(KEY_CURRENCY) == null) homeCurrency else null
+        val forHome: String? =
+            if (uri.getQueryParameter(KEY_ACCOUNTID) == null && uri.getQueryParameter(KEY_CURRENCY) == null) homeCurrency else null
 
         val group = enumValueOrDefault(uri.pathSegments[2], Grouping.NONE)
 
@@ -2048,7 +2050,11 @@ abstract class BaseTransactionProvider : ContentProvider() {
                 mergeCategory(it.getLong(0), it.getLong(1))
             }
         }
-        fun update(table: String, column: String = KEY_CATID, conflictAlgorithm: Int = CONFLICT_NONE) {
+        fun update(
+            table: String,
+            column: String = KEY_CATID,
+            conflictAlgorithm: Int = CONFLICT_NONE,
+        ) {
             update(
                 table,
                 ContentValues(1).apply {
@@ -2179,12 +2185,13 @@ abstract class BaseTransactionProvider : ContentProvider() {
                 """.trimMargin()
             )
             //parts are promoted to independence
-            val partUpdateCount = compileStatement("UPDATE $TABLE_TRANSACTIONS SET $KEY_PARENTID = null, $KEY_CR_STATUS = $crStatusSubSelect, $KEY_PAYEEID = $payeeIdSubSelect WHERE $KEY_PARENTID = ?").use {
-                it.bindString(1, uuid)
-                it.bindString(2, uuid)
-                it.bindLong(3, transactionId)
-                it.executeUpdateDelete()
-            }
+            val partUpdateCount =
+                compileStatement("UPDATE $TABLE_TRANSACTIONS SET $KEY_PARENTID = null, $KEY_CR_STATUS = $crStatusSubSelect, $KEY_PAYEEID = $payeeIdSubSelect WHERE $KEY_PARENTID = ?").use {
+                    it.bindString(1, uuid)
+                    it.bindString(2, uuid)
+                    it.bindLong(3, transactionId)
+                    it.executeUpdateDelete()
+                }
             //Change is recorded
             if (callerIsNotSyncAdapter) {
                 execSQL(
@@ -2260,25 +2267,44 @@ abstract class BaseTransactionProvider : ContentProvider() {
     val dynamicCurrenciesSelection
         get() = "$dynamicExchangeRatesDefault AND $KEY_CURRENCY != ?"
 
-    /**
-     * @param transactionId can be passed in as Long or String
-     */
     fun SupportSQLiteDatabase.insertOrReplaceEquivalentAmount(
         transactionId: Long,
         equivalentAmount: Long,
-    ) {
+    ) =
         // we do not use "Insert or replace" because we would receive insert trigger instead of
-        // update trigger and we would not be able to check if value has changed
-        val count = update(
+        // update trigger, and we would not be able to check if value has changed
+        update(
             TABLE_EQUIVALENT_AMOUNTS,
             ContentValues(1).apply {
                 put(KEY_EQUIVALENT_AMOUNT, equivalentAmount)
             }, "$KEY_TRANSACTIONID = ? AND $KEY_CURRENCY = ?", arrayOf(transactionId, homeCurrency)
-        )
-        if (count == 0) {
+        ).takeIf { it != 0 } ?: run {
             insertEquivalentAmount(transactionId, equivalentAmount)
+            1
+        }
+
+    fun SupportSQLiteDatabase.insertOrReplaceEquivalentAmount(
+        equivalentAmount: Long,
+        where: String,
+        whereArgs: Array<Any>,
+    ) = query(
+        table = TABLE_TRANSACTIONS,
+        columns = arrayOf(KEY_ROWID),
+        selection = where,
+        selectionArgs = whereArgs
+    ).use {
+        if (it.moveToFirst()) {
+            var count = 0
+            do {
+                count += insertOrReplaceEquivalentAmount(it.getLong(0), equivalentAmount)
+            } while (it.moveToNext())
+            count
+        } else {
+            report(IllegalArgumentException("selection = $where, args  = ${whereArgs.joinToString()} "))
+            0
         }
     }
+
 
     /**
      * also populates account exchange rates where missing and possible
@@ -2520,7 +2546,9 @@ abstract class BaseTransactionProvider : ContentProvider() {
 
     fun SupportSQLiteDatabase.deleteDebt(debId: String, where: String?, whereArgs: Array<String>?) =
         safeUpdateWithSealed(protectDebts = false) {
-            delete(TABLE_DEBTS,
-                "$KEY_ROWID = $debId${prefixAnd(where)}", whereArgs)
+            delete(
+                TABLE_DEBTS,
+                "$KEY_ROWID = $debId${prefixAnd(where)}", whereArgs
+            )
         }
 }
